#3108 Clean Cube Index Output report error on EMR 4.7

+++ extensions/engine-mr/src/main/java/io/kyligence/kap/engine/mr/steps/UpdateOutputDirStep.java
@@ -134,7 +134,9 @@ public class UpdateOutputDirStep extends AbstractExecutable {
             if (fs.exists(destChild) && fs.isDirectory(status.getPath()) && fs.isDirectory(destChild)) {
                 copyMergeDirs(fs, status.getPath(), destChild);
             } else {
-                fs.rename(status.getPath(), destChild);
+                if (fs.isFile(status.getPath()) && !fs.exists(destChild)) {
+                    fs.rename(status.getPath(), destChild);
+                }
             }
         }
     }
